There was discovered a logic, that the feature is created in the operation, after the operation can perform some actions for feature and after, outside, the property panel widgets are created and the feature is initialized by the control default values.
The problem with the Sketch creation line operation, which need to fill the property start point by the last point of another feature. The functionality the operation done in the startOperation() virtual method. After, the property panel throw down the values set.
The decision is to redesign the operation-widget_factory-property panel functionality in order to set the default values to the feature at the moment of a new feature creation.