refs #187: Property panel is in SALOME desktop
The property for widget's states is switched off in order to provide hidden property panel widget even if it was shown before the module deactivation.
There was an attemp to react on visibilityChanged of the property panel, but the current SALOME/Qt realization gives invalid painted area in the application on the last place of the panel.