Edit operation correction.
Edit and multi edit union finish, restart edit operation from sketch and edit operation only.
Do not perform the commit of edit operation from the module.
The edit operation is commited on mousePress(). This is the temporary modification until the selection(local, global) restore is not realized on setShape() for AIS object.