Deactivate an arc AIS preview object.
If the AIS exists immediatelly after the feature create, it is not deactivated because of Wokshop's realization.
The deactivation happens by the signal about the feature redisplay. The result on a signal about the feature create is display, but not the deactivate. In case of the AIS object exists immediatelly, the slot onFeatureCreated should be corrected to deactivate the AIS after display.