refs #240 - Selection of dimension for edit
the active widget is a selection control. The operation should not be committed on the mouse release.
Only if it is a last selected widget.
The disappearance of the property panel is corrected for the perpendicular constraint edition.
The case is call edit on a perpendicular constraint, select any line, it is applied as a first parameter, but the property panel is hidden.