Issue #351 constraint is not applied on preselected segments
The reason: lost selection of the circle center.
The decision: do not perform redisplay the AIS presentation in the AIS_InteractiveContext(displayer) if the shapes are equal. Here the GeomAPI realisation is necessary. The TopoDS_Shape realisation of the isEqual method is not enough because different TShapes instances are created and isEqual returns false.
A correction in the widget shape selector to avoid the distance presentation update by set an object(before set an attribute). In the previous version, the distance was jumped by selecting two points on lines. The reason is the distance calculation between point-object by setObject and after that point-point by setAttr