Issue #348 Validate sketch is disabled when point coordinates are set manually
It saves the clicked point, connects to the the activated signal of a widget and try initialize the widget with the point value.
There are some limitations:
Firstly, for the constraints distane we should not perform it because the fly out point is moved to the mouse, there is a blinking between calculated value and the clicked. It can be corrected if do not update the viewer during this. But, the lightes workaround is to do not apply clicked point to such type of operation.
Secondly, the mouse release point should be also saved in the clicked point for the mouse moving with Left button swithed. Without this, the actuvatedwidget value is moved to the previuosly clicked mouse but not the current one.