Issue #368 move constraint problem
The operation should be applyed if the apply state is true in the operation manager, otherwise, it is aborted.
The isOkEnabled, setOkEnabled methods should not be avaialabe in the property panel because this state is updated according to operation manager state by workshop object.
canAbortOperation is renamed to canStopOperation, because it is possible that the given operation is commited or aborted.