Construction elements are auxiliary entities:
a boolean flag in the sketch feature,
customize sketch feature presentation by the color change.

Important modification: it was decided to use non-obligatory xml constructions to set the no-focus policy for the attribute.
Otherwise during a segment-line creation, it is not enough to select two points, the check box control has a focus after them.