Fillet visualization problem: the first visualization is not correct.
The model correction to use the up-to-date number of composite features(it is increased by fillet execution)
Sketch manager should not filter the features and results, which are not related to the operation feature (fillet arc, coincidents, radius).
Constraint fillet is corrected: setInitialized() for a list attribute is not correct because currently the list combine this condition with a list size.
The recalculate of the center of an arc is not necessary because it happens in earlier code(azv)