An improvement to deselect a value in a shape selector control in the same way as it is realized in multi selector.
The reason is the plane deselection in the extrusion operation.

This leads to a modification in selection behavior e.g for sketch parallel constraint. After the first line is selected, the second control gets the focus, it is empty, so, the selection in the viewer is also empty. So, the first object is not selected anymore and the user can not see it.