A fix for activate/deactivate selection from destructor. It is possible, that selection mode is not correct by a widget activation.
Scenario: sketch, create a poligon, start parallel operation, select two lines, result - a create new parallel feature operation is started. The first control is active, but previous parallel constraint is selectable. This is wrong. If the next operation is translation, the parallel constraint selection leads to crash.
Reason: The activateShapeSselection(false) called in the destructor of widget shape selector is called after starting/activation of a widget in a new operation. So, it clears the previous activation of the current operation, deactivate reset it.