Result compsolid should change the display flag of sub-results itself.
WorkshopListener in displayObject() should ignore the compsolid result and wait when sub-solid results comes in selection.
onFeatureCreateMsg moving rows: 1.canBeDisplayed before setDisplayed, displayObject - this is a previous version. It is more correctly firstly change display state of the object and after try to display it in the viewer. It is possible that flag is switch on, but the object will be visualized during redisplay.