#1041 Sometimes when setting several distance constraints, the first input field is inactive
XGUI_PropertyPanel::activateNextWidget: the function is not finished yet, only setFocus is called. During this setFocus, the WidgetEditor leads to restart of the distance operation. When the loop returned back to the activateNextWidget, aNextWidget=0(sometimes or trash) and the empty widget is activated(after the distance operation has been already started with existed active first widget).