Simplification of the code: it removes focusNextPrevChild processing.
It is not enough to emit signal about focus change here because it can be changed by mouse click in another control. It is not processed in this method.
The same functionality(to store modified value on focus lost) is realized in ::deactivate() of the model widget and processEnter method of the widge.