#816 In multi-rotation, be able to put the total angle or the step angle (Rotation only)
Realization for Translation
Correction for right focus change after the end point in the create operation selection. Later, it could be corrected if there is an alternative method to validators->isCase() is implemented. It should answer which controls are visible. It is important here, where the same attributes can be used in both switch cases but we do not like to focus those, which are not currently being used.