#1108 Point coordinates is changed after click in the viewer
The problem was in myClickedPoint, it was not cleared and after click, the new line used the old clicked coordinates. The point and coincide last point of the previous line are moved to a start point of a new line.
There is a problem with this flag during Alt+TAB when creation of a line is active.
So, the decision is to remove this flag. The case to be checked: contour creation. The preview of created lines should not dragged to an origin. How it is corrected: point 2D widget do not initialize attribute by activating.