#1141 Part is not deleted
It was decided to perform Delete of Parts on the XGUI level. In the previous realization the RemoveFeature was used for this.
However, as the result of not loaded Parts do not exists and this feature are simple, it was decided with Mikhail, that this feature(Part) will be removed in the same way as other features. Later, it is possible, the RemoveFeature will be deleted from the application tool bar.