Issue #1111 Sketch invisible
It is not enough to create selector on initialize(). The problem still exists in the second study because it is not called there. So, it is decided to perform this in connectToStudy() of module.
An alternative is to do this in NewGEOM_DataModel before Mgr->load(). The problem is in this place because it causes redisplay signals flush. The viewer should be initialized before this case.