Adjust 'idler_plate.py' model to the new behavior
According to the general limitation of the selection mechanism, it is impossible to separate the coincident centers of circles/arcs if they are projected to another sketch. So, the loading of the geometric dump always selects the center of the latest edge. The 'idler_plate.py' test becomes invalid after #19990 while checking the dump to Python, because centers of arcs are taken into account now when loading the geometric dump.