Issue #20456: Another dump study fails
Implement additional check for user-defined name of a feature. This provides correct dump of names of features, which results have named after the feature name, but their default name has to be composed from the name of the concealed result. For example, in this case 'Face_1' feature conceals 'Pipe_1_1' result, so, the default name of face's result is 'Pipe_1_1', however, is was manually renamed to 'Face_1_1'.