Test1967: Update tolerance ratio for moving vs creating.
It was expected that the time to move a point was 1/10 the time
needed to create the sketch. This ratio is just to restrictive
and proved to be wrong as many time we get an assertion error
on some machine:

AssertionError: Time to move point 0.024923065677285194 is greater than expected 0.01865600571036339

We allow for 1/5 of the time now. This is not perfect nor good but
the previous ratio was boggus anyway. Maybe at some point we
will need to rewrite this test for something more stable in time.